separate power grid test runners
